refactor(editor): studio residual 분리
