+++ front/src/routes/Admin/editorStudioMetaModel.ts
@@ -12,6 +12,9 @@ import {
   stripThumbnailFocusFromUrl,
 } from "src/libs/thumbnailFocus"
 import type { PostVisibility } from "./editorStudioState"
+import { dedupeStrings, normalizeRecommendedTags } from "./editorStudioMetaModelHelpers"
+
+export { dedupeStrings, normalizeRecommendedTags } from "./editorStudioMetaModelHelpers"
 
 export type ParsedEditorMeta = {
   body: string
@@ -66,15 +69,6 @@ const INVISIBLE_CODE_PLACEHOLDER_REGEX = /[\u00A0\u200B-\u200D\u2060\uFEFF]/g
 
 export const PREVIEW_SUMMARY_MAX_LENGTH = 150
 export const PREVIEW_SUMMARY_MAX_CONTENT_LENGTH = 50_000
-
-export const dedupeStrings = (items: string[]) =>
-  Array.from(
-    new Set(
-      items
-        .map((item) => item.trim())
-        .filter(Boolean)
-    )
-  )
 
 const normalizeMetaItems = (raw: string): string[] => {
   const normalized = raw.trim().replace(/^\[|\]$/g, "")
@@ -198,23 +192,6 @@ export const normalizeSafePreviewThumbnailUrl = (raw: string): string => {
 
 export const makePreviewSummary = (content: string, maxLength = PREVIEW_SUMMARY_MAX_LENGTH) =>
   buildPreviewSummaryFromMarkdown(content, maxLength, "")
-
-export const normalizeRecommendedTags = (value: unknown, maxTags: number) => {
-  if (!Array.isArray(value)) return []
-  const map = new Map<string, string>()
-  value.forEach((item) => {
-    if (typeof item !== "string") return
-    const normalized = item.replace(/[\r\n]/g, " ").replace(/#/g, "").replace(/\s+/g, " ").trim()
-    if (!normalized) return
-    if (normalized.length < 2 || normalized.length > 24) return
-    if (!/[\p{L}\p{N}]/u.test(normalized)) return
-    if (normalized.toLowerCase() === "aside") return
-    const key = normalized.toLowerCase()
-    if (map.has(key) || map.size >= maxTags) return
-    map.set(key, normalized)
-  })
-  return Array.from(map.values())
-}
 
 export const resolveTagRecommendationErrorMessage = (error: unknown) => {
   const message = error instanceof Error ? error.message : String(error)
